+++ reindeer-macros/src/lib.rs
@@ -1,5 +1,6 @@
 mod entity_data;
 mod structure;
+mod relations;
 
 use entity_data::EntityData;
 use proc_macro::TokenStream;
